Properly failing test
This is an acceptance level test that runs the game in the same way that
the `ruby game_runner.rb` will run it.

The test expects that when you call `run` on the instance, it will print
out the board.

That the game quits isn't tested. Right now, it doesn't act like a game
looping for input and producing output. When it does, we will have more
tests that ensure it quits at the right moment.